backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
bitarray==2.9.2
certifi==2024.7.4
charset-normalizer==3.3.2
ckzg==1.0.2
click==8.1.7
cmake==3.30.1
cytoolz==0.12.3
Django==3.2.21
django-cors-headers==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
dlib==19.24.4
eth-account==0.11.2
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==1.0.1
eth-typing==4.4.0
eth-utils==4.1.1
eth_abi==5.1.0
face-recognition==1.3.0
face_recognition_models==0.3.0
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.7
jsonfield==3.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
multidict==6.0.5
numpy==1.26.0
opencv-python==4.10.0.84
parsimonious==0.10.0
pillow==10.4.0
protobuf==5.27.2
psycopg2-binary==2.9.9
pycryptodome==3.20.0
PyJWT==1.7.1
python-dotenv==1.0.1
pytz==2023.3.post1
pyunormalize==15.1.0
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rlp==4.0.1
rpds-py==0.19.1
setuptools==71.1.0
sqlparse==0.4.4
toolz==0.12.1
typing_extensions==4.7.1
urllib3==2.2.2
web3==6.20.1
websockets==12.0
wheel==0.43.0
yarl==1.9.4
